refactor: move type to type dir
