--- src/app/[locale]/history/request-history.type.ts
@@ -1,16 +1,0 @@
-export type HttpMethod = 'GET' | 'POST' | 'PUT' | 'DELETE' | 'PATCH' | 'HEAD' | 'OPTIONS';
-
-export interface RequestHistoryItem {
-  id: string;
-  url: string;
-  method: HttpMethod;
-
-  timestamp: string;
-  durationMs: number;
-  statusCode: number;
-
-  requestSize: number;
-  responseSize: number;
-
-  errorDetails?: string;
-}
